requirements.txt
Django>=3.2
PyPDF2
python-docx
nltk
PyMuPDF
Pillow
pytesseract 
